Creates generic CapacityHelper rather than having the WorkspaceHelper rely on ReservationHelper
